+++ src/main/java/online/kingdomkeys/kingdomkeys/client/ClientUtils.java
@@ -770,6 +770,9 @@ public class ClientUtils {
         return playerData;
     }
 
+    /**
+     * Used in the KO system so it doesn't rotate
+     */
     public static void renderNameTag(LivingEntityRenderer<AbstractClientPlayer, PlayerModel<AbstractClientPlayer>> renderer, LivingEntity entity, String displayName, PoseStack poseStack, MultiBufferSource bufferSource, int packedLight, float partialTick) {
         EntityRenderDispatcher dispatcher = Minecraft.getInstance().getEntityRenderDispatcher();